flask_container/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv   
SQLAlchemy
pyspark
plotly
pymysql
numpy